Refactoring + some javadoc
